+++ packages/atom.io/__tests__/realtime-ipc/game-servers.test.tsx
@@ -1,16 +1,8 @@
-import path from "path"
 import { act } from "@testing-library/react"
-import {
-	actUponStore,
-	arbitrary,
-	findInStore,
-	getFromStore,
-} from "atom.io/internal"
-import type { Json } from "atom.io/json"
-import * as RTS from "atom.io/realtime-server"
 import * as RTTest from "atom.io/realtime-testing"
 import { BrowserGame } from "./BrowserGame"
 import { DatabaseManager } from "./database.node"
+import { SystemServer } from "./system-server.node"
 
 const dbManager = new DatabaseManager()
 
@@ -36,82 +28,7 @@ afterAll(async () => {
 describe(`multi-process realtime server`, () => {
 	const scenario = () => {
 		const { server, client, teardown } = RTTest.singleClient({
-			server: ({ socket, silo: { store } }) => {
-				const userKeyState = findInStore(
-					RTS.usersOfSockets.states.userKeyOfSocket,
-					socket.id,
-					store,
-				)
-				const userKey = getFromStore(userKeyState, store)
-
-				const exposeMutable = RTS.realtimeMutableProvider({ socket, store })
-				exposeMutable(RTS.roomIndex)
-
-				const exposeMutableFamily = RTS.realtimeMutableFamilyProvider({
-					socket,
-					store,
-				})
-				exposeMutableFamily(
-					RTS.usersOfSockets.core.findRelatedKeysState,
-					RTS.socketIndex,
-				)
-				exposeMutableFamily(
-					RTS.usersInRooms.core.findRelatedKeysState,
-					RTS.userIndex,
-				)
-				socket.on(`create-room`, async (roomId) => {
-					actUponStore(RTS.createRoomTX, arbitrary(), store)(roomId, `bun`, [
-						path.join(__dirname, `game-instance.bun.ts`),
-					])
-				})
-
-				let forwardToRoom:
-					| ((event: string, ...args: unknown[]) => void)
-					| undefined
-				const queuedEventsForRoom = new Map<string, unknown[]>()
-				socket.on(`join-room`, async (roomId) => {
-					if (!userKey) {
-						throw new Error(`User not found`)
-					}
-					forwardToRoom = (event: string, ...args: unknown[]) => {
-						queuedEventsForRoom.set(event, args)
-					}
-					actUponStore(RTS.joinRoomTX, arbitrary(), store)(roomId, userKey, 0)
-
-					const roomState = findInStore(RTS.roomSelectors, roomId, store)
-					const room = await getFromStore(roomState, store)
-					const roomSocket = new RTS.ChildSocket(room)
-					roomSocket.emit(`setup-relay`, userKey)
-
-					forwardToRoom = (event: string, ...args: Json.Array) => {
-						roomSocket.emit(event, ...args)
-					}
-					for (const [event, args] of queuedEventsForRoom.entries()) {
-						forwardToRoom(`relay:${userKey}`, event, ...args)
-					}
-					queuedEventsForRoom.clear()
-
-					roomSocket.onAny((...data) => socket.emit(...data))
-
-					room.stderr.on(`data`, (buf) => {
-						const err = buf.toString()
-						console.error(`❌ ${roomId} [${room.pid}]\n${err}`)
-					})
-
-					room.on(`close`, (code) => {
-						console.log(`${roomId} exited with code ${code}`)
-						socket.emit(`room-close`, roomId, code)
-					})
-				})
-
-				socket.onAny((event, ...args) => {
-					console.log(`🛰 `, userKey, event, ...args)
-					forwardToRoom?.(event, ...args)
-				})
-				socket.onAnyOutgoing((event, ...args) => {
-					console.log(`🛰  >>`, userKey, event, ...args)
-				})
-			},
+			server: SystemServer,
 			client: BrowserGame,
 		})
 
@@ -119,7 +36,7 @@ describe(`multi-process realtime server`, () => {
 	}
 
 	it(`runs several instances of the same server`, async () => {
-		const { client, server, teardown } = scenario()
+		const { client, teardown } = scenario()
 		const app = client.init()
 		const createRoomButton = await app.renderResult.findByTestId(`create-room`)
 		act(() => createRoomButton.click())
